services/coder-agent/pyproject.toml
[project]
name = "coder-agent"
version = "0.1.0"
description = "LangChain-based coder agent — POC on GCP Cloud Run (plain chat loop)"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "httpx>=0.27",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "google-auth>=2.35",
    "google-cloud-storage>=2.18",
    "python-json-logger>=2.0",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.7",
    "mypy>=1.13",
    "types-requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coder_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"
markers = [
    "slow: tests that hit real networks or take > 5s",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]  # handled by formatter

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["google.auth.*", "google.oauth2.*"]
ignore_missing_imports = true
